+++ experimental/packages/opentelemetry-instrumentation-http/src/utils.ts
@@ -295,6 +295,26 @@ export const isValidOptionsType = (options: unknown): boolean => {
   return type === 'string' || (type === 'object' && !Array.isArray(options));
 };
 
+export const extractHostnameAndPort = (
+  requestOptions: Pick<ParsedRequestOptions, 'hostname' | 'host' | 'port' | 'protocol'>
+): { hostname: string, port: number | string } => {
+  if (requestOptions.hostname && requestOptions.port) {
+    return {hostname: requestOptions.hostname, port: requestOptions.port};
+  }
+  const matches = requestOptions.host?.match(/^([^:/ ]+)(:\d{1,5})?/) || null;
+  const hostname = requestOptions.hostname || (matches === null ? 'localhost' : matches[1]);
+  let port = requestOptions.port;
+  if (!port) {
+    if (matches && matches[2]) {
+      // remove the leading ":". The extracted port would be something like ":8080"
+      port = matches[2].substring(1);
+    } else {
+      port = requestOptions.protocol === 'https:' ? '443' : '80';
+    }
+  }
+  return {hostname, port};
+};
+
 /**
  * Returns outgoing request attributes scoped to the options passed to the request
  * @param {ParsedRequestOptions} requestOptions the same options used to make the request
@@ -302,13 +322,10 @@ export const isValidOptionsType = (options: unknown): boolean => {
  */
 export const getOutgoingRequestAttributes = (
   requestOptions: ParsedRequestOptions,
-  options: { component: string; hostname: string; hookAttributes?: SpanAttributes }
+  options: { component: string; hostname: string; port: string | number, hookAttributes?: SpanAttributes }
 ): SpanAttributes => {
-  const host = requestOptions.host;
-  const hostname =
-    requestOptions.hostname ||
-    host?.replace(/^(.*)(:[0-9]{1,5})/, '$1') ||
-    'localhost';
+  const hostname = options.hostname;
+  const port = options.port;
   const requestMethod = requestOptions.method;
   const method = requestMethod ? requestMethod.toUpperCase() : 'GET';
   const headers = requestOptions.headers || {};
@@ -322,6 +339,7 @@ export const getOutgoingRequestAttributes = (
     [SemanticAttributes.HTTP_METHOD]: method,
     [SemanticAttributes.HTTP_TARGET]: requestOptions.path || '/',
     [SemanticAttributes.NET_PEER_NAME]: hostname,
+    [SemanticAttributes.HTTP_HOST]: requestOptions.headers?.host ?? `${hostname}:${port}`,
   };
 
   if (userAgent !== undefined) {
@@ -354,14 +372,12 @@ export const getAttributesFromHttpKind = (kind?: string): SpanAttributes => {
  */
 export const getOutgoingRequestAttributesOnResponse = (
   response: IncomingMessage,
-  options: { hostname: string }
 ): SpanAttributes => {
   const { statusCode, statusMessage, httpVersion, socket } = response;
   const { remoteAddress, remotePort } = socket;
   const attributes: SpanAttributes = {
     [SemanticAttributes.NET_PEER_IP]: remoteAddress,
     [SemanticAttributes.NET_PEER_PORT]: remotePort,
-    [SemanticAttributes.HTTP_HOST]: `${options.hostname}:${remotePort}`,
   };
   setResponseContentLengthAttribute(response, attributes);