+++ packages/opentelemetry-core/src/trace/sampler/TraceIdRatioBasedSampler.ts
@@ -21,7 +21,10 @@ import {
   isValidTraceId,
 } from '@opentelemetry/api';
 
-/** Sampler that samples a given fraction of traces based of trace id deterministically. */
+/**
+ * @deprecated Use the one defined in @opentelemetry/sdk-trace-base instead.
+ * Sampler that samples a given fraction of traces based of trace id deterministically.
+ */
 export class TraceIdRatioBasedSampler implements Sampler {
   private _upperBound: number;